Add files via upload
The final year project titled “Potential Study of Medium Hydropower Project of Nepal: A Case Study of Kabeli-A HEP” aims to evaluate the feasibility of the Kabeli-A Hydroelectric Project located in Panchthar and Taplejung districts. The study focuses on hydrological analysis, hydraulic design of major components like intake, weir, tunnel, and penstock, as well as economic evaluation of the project. Using GIS tools and engineering guidelines, the project assesses water flow, energy potential, structural stability, and cost effectiveness to provide a comprehensive overview of the project's viability.